fix: dependecy cycle + computed side effect
