+++ app/.eslintrc.js
@@ -18,6 +18,6 @@ module.exports = {
     'no-underscore-dangle': 'off',
     'no-undef': 'off',
     'no-plusplus': 'off',
-    "max-len": 'off',
+    'max-len': 'off',
   },
 };
