+++ app/src/shTimeFunction.js
@@ -1,0 +1,18 @@
+import { DateTime } from 'luxon';
+
+export default function shTimeFunction(date) {
+  let tempDate = date;
+  if (!Array.isArray(tempDate)) {
+    tempDate = [tempDate];
+  }
+  const dateObj = DateTime.fromISO(tempDate[0]);
+  const defaultFormat = "yyyy-MM-dd'T'HH:mm:ss";
+  const alternativeFormat = 'yyyy-MM-dd';
+  if (dateObj.second === 0 && dateObj.hour === 0 && dateObj.minute === 0) {
+    // if only day input, format as an interval to next day
+    const nextDay = dateObj.plus({ days: 1 });
+    return `${dateObj.toFormat(alternativeFormat)}/${nextDay.toFormat(alternativeFormat)}`;
+  }
+  // otherwise return single date with full format
+  return `${dateObj.toFormat(defaultFormat)}/${dateObj.toFormat(defaultFormat)}`;
+}
